Fix remainingBytes type

+++ src/ESP32_VS1053_Stream.h
@@ -90,7 +90,7 @@ private:
 
     unsigned long _startMute = 0;
     size_t _offset = 0;
-    size_t _remainingBytes = 0;
+    int32_t _remainingBytes = 0;
     size_t _bytesLeftInChunk = 0;
     int32_t _metaDataStart = 0;
     int32_t _musicDataPosition = 0;
